Add database object for accesing and storing PDGuard fields
Configure application to setup an in memory database using
Anorm. Used to store PDGuard fields associated with a guardian
user in order to used by current application to issue PDGuard requests.
Moreover,a SQL script with the creation of the corresponding table
and the insertion of initial test data is added.